+++ je_auto_control/utils/perceptual_diff/__init__.py
@@ -1,0 +1,6 @@
+"""Perceptual (YIQ) image diff with anti-alias edge suppression."""
+from je_auto_control.utils.perceptual_diff.perceptual_diff import (
+    PerceptualDiffResult, assert_perceptual, perceptual_diff,
+)
+
+__all__ = ["PerceptualDiffResult", "assert_perceptual", "perceptual_diff"]